+++ KioChess/Engine/Strategies/Models/AspirationModel.cs
@@ -4,7 +4,7 @@ namespace Engine.Strategies.Models;
 
 public class AspirationModel
 {
-    public short Window { get; set; }
+    public int Window { get; set; }
     public sbyte Depth { get; set; }
     public MemoryStrategyBase Strategy { get; set; }
 }
